Moved more_plot.py to scripts/
